requirements.txt
python-bcb==0.3.3
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2
scikit-learn==1.5.0
xgboost==2.0.3
shap==0.45.1
plotly==5.22.0
streamlit==1.35.0
openai==1.30.0
langchain==0.2.3
langchain-openai==0.1.8
python-dotenv==1.0.1
joblib==1.4.2
